requirements.txt
flask
transformers
requests
nltk
python-dotenv
tensorflow  # Or replace with 'torch' if using PyTorch
tf-keras